Move the public files into the data directory.
`data_path` is userland and so the group images should live there. There
should also be public files for techcity (like a default `robots.txt`)
that should remain in the package. This change splits those concepts out
in the build step so that both areas are copied into the final output.
The userland data will override any files from techcity.